fermant: increase range for lincom coefficients
g(3,3) shows that 16-bit coefficients are not enough.

Additionally, remove vestigial support for separately storing size of
lincom in terms which was masking a bug that we need to allow for
alignment in the term_t structure.

+++ fermant/lincom.h
@@ -7,7 +7,7 @@
 #include "int.h"
 #include "num.h"
 
-typedef signed short lincomz_t;
+typedef signed int lincomz_t;
 /* notionally:
     typedef struct {
         lincomz_t value[nv + 1];
